add fake "-verbose" argument to QEMU
Problem: the Eclipse plugin always passes an argument "--verbose" to QEMU. The
normal QEMU does not understand this argument -- it is a special argument added
to the fork created by the gnuarmeclipse people (github.com/gnuarmeclipse/qemu)

The help files and such seem to be automatically generated, so we also add some
fake entries in qemu-options.def & qemu-options.hx.

In vl.c there is a function lookup_opt() which checks that all arguments are
legit (do they exist and, if they need params, does that exist).