changed configure to disable AVX optimizations
configure tries to detect if QEMU can use AVX instructions to improve
performance by compiling a test application using those instructions.

But it checks if this machine can use AVX instructions - which might
not be the case for the other machines this runs on.

The default is "don't use AVX", the check could potentitally set it to
"do use AVX"; simply commenting that check out always disables them